refactor: broke state classes into separate files
