+++ backend/consensus/helpers/consensus_algorithm.py
@@ -1,0 +1,1238 @@
+import os
+import asyncio
+import traceback
+from typing import Callable, List
+import time
+import threading
+
+from sqlalchemy.orm import Session
+from backend.consensus.vrf import get_validators_for_transaction
+from backend.database_handler.chain_snapshot import ChainSnapshot
+from backend.database_handler.contract_snapshot import ContractSnapshot
+from backend.database_handler.transactions_processor import (
+    TransactionsProcessor,
+    TransactionStatus,
+)
+from backend.database_handler.accounts_manager import AccountsManager
+from backend.database_handler.types import ConsensusData
+from backend.domain.types import (
+    Transaction,
+)
+from backend.node.base import Node
+from backend.node.types import (
+    ExecutionMode,
+    Receipt,
+)
+from backend.protocol_rpc.message_handler.base import MessageHandler
+from backend.protocol_rpc.message_handler.types import (
+    LogEvent,
+    EventType,
+    EventScope,
+)
+from backend.consensus.base import (
+    chain_snapshot_factory,
+    transactions_processor_factory,
+    accounts_manager_factory,
+    contract_snapshot_factory,
+    node_factory,
+    DEFAULT_CONSENSUS_SLEEP_TIME,
+)
+from backend.consensus.helpers.transaction_context import TransactionContext
+from backend.consensus.states.pending_state import PendingState
+from backend.consensus.states.committing_state import CommittingState
+from backend.consensus.states.finalizing_state import FinalizingState
+
+
+class ConsensusAlgorithm:
+    """
+    Class representing the consensus algorithm.
+
+    Attributes:
+        get_session (Callable[[], Session]): Function to get a database session.
+        msg_handler (MessageHandler): Handler for messaging.
+        pending_queues (dict[str, asyncio.Queue]): Dictionary of pending_queues for transactions.
+        finality_window_time (int): Time in seconds for the finality window.
+        consensus_sleep_time (int): Time in seconds for the consensus sleep time.
+    """
+
+    def __init__(
+        self,
+        get_session: Callable[[], Session],
+        msg_handler: MessageHandler,
+    ):
+        """
+        Initialize the ConsensusAlgorithm.
+
+        Args:
+            get_session (Callable[[], Session]): Function to get a database session.
+            msg_handler (MessageHandler): Handler for messaging.
+        """
+        self.get_session = get_session
+        self.msg_handler = msg_handler
+        self.pending_queues: dict[str, asyncio.Queue] = {}
+        self.finality_window_time = int(os.getenv("VITE_FINALITY_WINDOW"))
+        self.consensus_sleep_time = DEFAULT_CONSENSUS_SLEEP_TIME
+        self.pending_queue_stop_events: dict[str, asyncio.Event] = (
+            {}
+        )  # Events to stop tasks for each pending queue
+        self.pending_queue_task_running: dict[str, bool] = (
+            {}
+        )  # Track running state for each pending queue
+
+    def run_crawl_snapshot_loop(
+        self,
+        chain_snapshot_factory: Callable[
+            [Session], ChainSnapshot
+        ] = chain_snapshot_factory,
+        transactions_processor_factory: Callable[
+            [Session], TransactionsProcessor
+        ] = transactions_processor_factory,
+        stop_event: threading.Event = threading.Event(),
+    ):
+        """
+        Run the loop to crawl snapshots.
+
+        Args:
+            chain_snapshot_factory (Callable[[Session], ChainSnapshot]): Creates snapshots of the blockchain state at specific points in time.
+            transactions_processor_factory (Callable[[Session], TransactionsProcessor]): Creates processors to modify transactions.
+            stop_event (threading.Event): Control signal to terminate the loop.
+        """
+        # Create a new event loop for crawling snapshots
+        loop = asyncio.new_event_loop()
+        asyncio.set_event_loop(loop)
+        loop.run_until_complete(
+            self._crawl_snapshot(
+                chain_snapshot_factory, transactions_processor_factory, stop_event
+            )
+        )
+        loop.close()
+
+    async def _crawl_snapshot(
+        self,
+        chain_snapshot_factory: Callable[[Session], ChainSnapshot],
+        transactions_processor_factory: Callable[[Session], TransactionsProcessor],
+        stop_event: threading.Event,
+    ):
+        """
+        Crawl snapshots and process pending transactions.
+
+        Args:
+            chain_snapshot_factory (Callable[[Session], ChainSnapshot]): Creates snapshots of the blockchain state at specific points in time.
+            transactions_processor_factory (Callable[[Session], TransactionsProcessor]): Creates processors to modify transactions.
+            stop_event (threading.Event): Control signal to terminate the loop.
+        """
+        while not stop_event.is_set():
+            with self.get_session() as session:
+                chain_snapshot = chain_snapshot_factory(session)
+                transactions_processor = transactions_processor_factory(session)
+                pending_transactions = chain_snapshot.get_pending_transactions()
+                for transaction in pending_transactions:
+                    transaction = Transaction.from_dict(transaction)
+                    address = transaction.to_address
+
+                    # Initialize queue and stop event for the address if not present
+                    if address not in self.pending_queues:
+                        self.pending_queues[address] = asyncio.Queue()
+
+                    if address not in self.pending_queue_stop_events:
+                        self.pending_queue_stop_events[address] = asyncio.Event()
+
+                    # Only add to the queue if the stop event is not set
+                    if not self.pending_queue_stop_events[address].is_set():
+                        await self.pending_queues[address].put(transaction)
+
+                        # Set the transaction as activated so it is not added to the queue again
+                        ConsensusAlgorithm.dispatch_transaction_status_update(
+                            transactions_processor,
+                            transaction.hash,
+                            TransactionStatus.ACTIVATED,
+                            self.msg_handler,
+                        )
+
+            await asyncio.sleep(self.consensus_sleep_time)
+
+    def run_process_pending_transactions_loop(
+        self,
+        chain_snapshot_factory: Callable[
+            [Session], ChainSnapshot
+        ] = chain_snapshot_factory,
+        transactions_processor_factory: Callable[
+            [Session], TransactionsProcessor
+        ] = transactions_processor_factory,
+        accounts_manager_factory: Callable[
+            [Session], AccountsManager
+        ] = accounts_manager_factory,
+        contract_snapshot_factory: Callable[
+            [str, Session, Transaction], ContractSnapshot
+        ] = contract_snapshot_factory,
+        node_factory: Callable[
+            [
+                dict,
+                ExecutionMode,
+                ContractSnapshot,
+                Receipt | None,
+                MessageHandler,
+                Callable[[str], ContractSnapshot],
+            ],
+            Node,
+        ] = node_factory,
+        stop_event: threading.Event = threading.Event(),
+    ):
+        """
+        Run the process pending transactions loop.
+
+        Args:
+            chain_snapshot_factory (Callable[[Session], ChainSnapshot]): Creates snapshots of the blockchain state at specific points in time.
+            transactions_processor_factory (Callable[[Session], TransactionsProcessor]): Creates processors to modify transactions.
+            accounts_manager_factory (Callable[[Session], AccountsManager]): Creates managers to handle account state.
+            contract_snapshot_factory (Callable[[str, Session, Transaction], ContractSnapshot]): Creates snapshots of contract states.
+            node_factory (Callable[[dict, ExecutionMode, ContractSnapshot, Receipt | None, MessageHandler, Callable[[str], ContractSnapshot]], Node]): Creates node instances that can execute contracts and process transactions.
+            stop_event (threading.Event): Control signal to terminate the pending transactions process.
+        """
+        # Create a new event loop for running the processing of pending transactions
+        loop = asyncio.new_event_loop()
+        asyncio.set_event_loop(loop)
+        loop.run_until_complete(
+            self._process_pending_transactions(
+                chain_snapshot_factory,
+                transactions_processor_factory,
+                accounts_manager_factory,
+                contract_snapshot_factory,
+                node_factory,
+                stop_event,
+            )
+        )
+        loop.close()
+
+    async def _process_pending_transactions(
+        self,
+        chain_snapshot_factory: Callable[[Session], ChainSnapshot],
+        transactions_processor_factory: Callable[[Session], TransactionsProcessor],
+        accounts_manager_factory: Callable[[Session], AccountsManager],
+        contract_snapshot_factory: Callable[
+            [str, Session, Transaction], ContractSnapshot
+        ],
+        node_factory: Callable[
+            [
+                dict,
+                ExecutionMode,
+                ContractSnapshot,
+                Receipt | None,
+                MessageHandler,
+                Callable[[str], ContractSnapshot],
+            ],
+            Node,
+        ],
+        stop_event: threading.Event,
+    ):
+        """
+        Process pending transactions.
+
+        Args:
+            chain_snapshot_factory (Callable[[Session], ChainSnapshot]): Creates snapshots of the blockchain state at specific points in time.
+            transactions_processor_factory (Callable[[Session], TransactionsProcessor]): Creates processors to modify transactions.
+            accounts_manager_factory (Callable[[Session], AccountsManager]): Creates managers to handle account state.
+            contract_snapshot_factory (Callable[[str, Session, Transaction], ContractSnapshot]): Creates snapshots of contract states.
+            node_factory (Callable[[dict, ExecutionMode, ContractSnapshot, Receipt | None, MessageHandler, Callable[[str], ContractSnapshot]], Node]): Creates node instances that can execute contracts and process transactions.
+            stop_event (threading.Event): Control signal to terminate the pending transactions process.
+        """
+        # Set a new event loop for the processing of pending transactions
+        asyncio.set_event_loop(asyncio.new_event_loop())
+        # Note: ollama uses GPU resources and webrequest aka selenium uses RAM
+        # TODO: Consider using async sessions to avoid blocking the current thread
+        while not stop_event.is_set():
+            try:
+                async with asyncio.TaskGroup() as tg:
+                    for queue_address, queue in self.pending_queues.items():
+                        if (
+                            not queue.empty()
+                            and not self.pending_queue_stop_events.get(
+                                queue_address, asyncio.Event()
+                            ).is_set()
+                        ):
+                            # Sessions cannot be shared between coroutines; create a new session for each coroutine
+                            # Reference: https://docs.sqlalchemy.org/en/20/orm/session_basics.html#is-the-session-thread-safe-is-asyncsession-safe-to-share-in-concurrent-tasks
+                            self.pending_queue_task_running[queue_address] = True
+                            transaction: Transaction = await queue.get()
+                            with self.get_session() as session:
+
+                                async def exec_transaction_with_session_handling(
+                                    session: Session,
+                                    transaction: Transaction,
+                                    queue_address: str,
+                                ):
+                                    transactions_processor = (
+                                        transactions_processor_factory(session)
+                                    )
+                                    await self.exec_transaction(
+                                        transaction,
+                                        transactions_processor,
+                                        chain_snapshot_factory(session),
+                                        accounts_manager_factory(session),
+                                        lambda contract_address: contract_snapshot_factory(
+                                            contract_address, session, transaction
+                                        ),
+                                        node_factory,
+                                    )
+                                    session.commit()
+                                    self.pending_queue_task_running[queue_address] = (
+                                        False
+                                    )
+
+                            tg.create_task(
+                                exec_transaction_with_session_handling(
+                                    session, transaction, queue_address
+                                )
+                            )
+
+            except Exception as e:
+                print("Error running consensus", e)
+                print(traceback.format_exc())
+            finally:
+                for queue_address in self.pending_queues:
+                    self.pending_queue_task_running[queue_address] = False
+            await asyncio.sleep(self.consensus_sleep_time)
+
+    def is_pending_queue_task_running(self, address: str):
+        """
+        Check if a task for a specific pending queue is currently running.
+        """
+        return self.pending_queue_task_running.get(address, False)
+
+    def stop_pending_queue_task(self, address: str):
+        """
+        Signal the task for a specific pending queue to stop.
+        """
+        if address in self.pending_queues:
+            if address not in self.pending_queue_stop_events:
+                self.pending_queue_stop_events[address] = asyncio.Event()
+            self.pending_queue_stop_events[address].set()
+
+    def start_pending_queue_task(self, address: str):
+        """
+        Allow the task for a specific pending queue to start.
+        """
+        if address in self.pending_queue_stop_events:
+            self.pending_queue_stop_events[address].clear()
+
+    async def exec_transaction(
+        self,
+        transaction: Transaction,
+        transactions_processor: TransactionsProcessor,
+        chain_snapshot: ChainSnapshot,
+        accounts_manager: AccountsManager,
+        contract_snapshot_factory: Callable[[str], ContractSnapshot],
+        node_factory: Callable[
+            [
+                dict,
+                ExecutionMode,
+                ContractSnapshot,
+                Receipt | None,
+                MessageHandler,
+                Callable[[str], ContractSnapshot],
+            ],
+            Node,
+        ],
+    ):
+        """
+        Execute a transaction.
+
+        Args:
+            transaction (Transaction): The transaction to execute.
+            transactions_processor (TransactionsProcessor): Instance responsible for handling transaction operations within the database.
+            chain_snapshot (ChainSnapshot): Snapshot of the chain state.
+            accounts_manager (AccountsManager): Manager for accounts.
+            contract_snapshot_factory (Callable[[str], ContractSnapshot]): Factory function to create contract snapshots.
+            node_factory (Callable[[dict, ExecutionMode, ContractSnapshot, Receipt | None, MessageHandler, Callable[[str], ContractSnapshot]], Node]): Factory function to create nodes.
+        """
+        # Create initial state context for the transaction
+        context = TransactionContext(
+            transaction=transaction,
+            transactions_processor=transactions_processor,
+            chain_snapshot=chain_snapshot,
+            accounts_manager=accounts_manager,
+            contract_snapshot_factory=contract_snapshot_factory,
+            node_factory=node_factory,
+            msg_handler=self.msg_handler,
+        )
+
+        # Begin state transitions starting from PendingState
+        state = PendingState()
+        while True:
+            next_state = await state.handle(context)
+            if next_state is None:
+                break
+            state = next_state
+
+    @staticmethod
+    def dispatch_transaction_status_update(
+        transactions_processor: TransactionsProcessor,
+        transaction_hash: str,
+        new_status: TransactionStatus,
+        msg_handler: MessageHandler,
+    ):
+        """
+        Dispatch a transaction status update.
+
+        Args:
+            transactions_processor (TransactionsProcessor): Instance responsible for handling transaction operations within the database.
+            transaction_hash (str): Hash of the transaction.
+            new_status (TransactionStatus): New status of the transaction.
+            msg_handler (MessageHandler): Handler for messaging.
+        """
+        # Update the transaction status in the transactions processor
+        transactions_processor.update_transaction_status(transaction_hash, new_status)
+
+        # Send a message indicating the transaction status update
+        msg_handler.send_message(
+            LogEvent(
+                "transaction_status_updated",
+                EventType.INFO,
+                EventScope.CONSENSUS,
+                f"{str(new_status.value)} {str(transaction_hash)}",
+                {
+                    "hash": str(transaction_hash),
+                    "new_status": str(new_status.value),
+                },
+                transaction_hash=transaction_hash,
+            )
+        )
+
+    @staticmethod
+    def execute_transfer(
+        transaction: Transaction,
+        transactions_processor: TransactionsProcessor,
+        accounts_manager: AccountsManager,
+        msg_handler: MessageHandler,
+    ):
+        """
+        Executes a native token transfer between Externally Owned Accounts (EOAs).
+
+        This function handles the transfer of native tokens from one EOA to another.
+        It updates the balances of both the sender and recipient accounts, and
+        manages the transaction status throughout the process.
+
+        Args:
+            transaction (dict): The transaction details including from_address, to_address, and value.
+            transactions_processor (TransactionsProcessor): Instance responsible for handling transaction operations within the database.
+            accounts_manager (AccountsManager): Manager to handle account balance updates.
+        """
+
+        # Check if the transaction is a fund_account call
+        if not transaction.from_address is None:
+            # Get the balance of the sender account
+            from_balance = accounts_manager.get_account_balance(
+                transaction.from_address
+            )
+
+            # Check if the sender has enough balance
+            if from_balance < transaction.value:
+                # Set the transaction status to UNDETERMINED if balance is insufficient
+                ConsensusAlgorithm.dispatch_transaction_status_update(
+                    transactions_processor,
+                    transaction.hash,
+                    TransactionStatus.UNDETERMINED,
+                    msg_handler,
+                )
+
+                transactions_processor.create_rollup_transaction(transaction.hash)
+                return
+
+            # Update the balance of the sender account
+            accounts_manager.update_account_balance(
+                transaction.from_address, from_balance - transaction.value
+            )
+
+        # Check if the transaction is a burn call
+        if not transaction.to_address is None:
+            # Get the balance of the recipient account
+            to_balance = accounts_manager.get_account_balance(transaction.to_address)
+
+            # Update the balance of the recipient account
+            accounts_manager.update_account_balance(
+                transaction.to_address, to_balance + transaction.value
+            )
+
+        # Dispatch a transaction status update to FINALIZED
+        ConsensusAlgorithm.dispatch_transaction_status_update(
+            transactions_processor,
+            transaction.hash,
+            TransactionStatus.FINALIZED,
+            msg_handler,
+        )
+
+        transactions_processor.create_rollup_transaction(transaction.hash)
+
+    def run_appeal_window_loop(
+        self,
+        chain_snapshot_factory: Callable[
+            [Session], ChainSnapshot
+        ] = chain_snapshot_factory,
+        transactions_processor_factory: Callable[
+            [Session], TransactionsProcessor
+        ] = transactions_processor_factory,
+        accounts_manager_factory: Callable[
+            [Session], AccountsManager
+        ] = accounts_manager_factory,
+        contract_snapshot_factory: Callable[
+            [str, Session, Transaction], ContractSnapshot
+        ] = contract_snapshot_factory,
+        node_factory: Callable[
+            [
+                dict,
+                ExecutionMode,
+                ContractSnapshot,
+                Receipt | None,
+                MessageHandler,
+                Callable[[str], ContractSnapshot],
+            ],
+            Node,
+        ] = node_factory,
+        stop_event: threading.Event = threading.Event(),
+    ):
+        """
+        Run the loop to handle the appeal window.
+
+        Args:
+            chain_snapshot_factory (Callable[[Session], ChainSnapshot]): Creates snapshots of the blockchain state at specific points in time.
+            transactions_processor_factory (Callable[[Session], TransactionsProcessor]): Creates processors to modify transactions.
+            accounts_manager_factory (Callable[[Session], AccountsManager]): Creates managers to handle account state.
+            contract_snapshot_factory (Callable[[str, Session, Transaction], ContractSnapshot]): Creates snapshots of contract states.
+            node_factory (Callable[[dict, ExecutionMode, ContractSnapshot, Receipt | None, MessageHandler, Callable[[str], ContractSnapshot]], Node]): Creates node instances that can execute contracts and process transactions.
+            stop_event (threading.Event): Control signal to terminate the appeal window process.
+        """
+        # Create a new event loop for running the appeal window
+        loop = asyncio.new_event_loop()
+        asyncio.set_event_loop(loop)
+        loop.run_until_complete(
+            self._appeal_window(
+                chain_snapshot_factory,
+                transactions_processor_factory,
+                accounts_manager_factory,
+                contract_snapshot_factory,
+                node_factory,
+                stop_event,
+            )
+        )
+        loop.close()
+
+    async def _appeal_window(
+        self,
+        chain_snapshot_factory: Callable[[Session], ChainSnapshot],
+        transactions_processor_factory: Callable[[Session], TransactionsProcessor],
+        accounts_manager_factory: Callable[[Session], AccountsManager],
+        contract_snapshot_factory: Callable[
+            [str, Session, Transaction], ContractSnapshot
+        ],
+        node_factory: Callable[
+            [
+                dict,
+                ExecutionMode,
+                ContractSnapshot,
+                Receipt | None,
+                MessageHandler,
+                Callable[[str], ContractSnapshot],
+            ],
+            Node,
+        ],
+        stop_event: threading.Event,
+    ):
+        """
+        Handle the appeal window for transactions, during which EOAs can challenge transaction results.
+
+        Args:
+            chain_snapshot_factory (Callable[[Session], ChainSnapshot]): Creates snapshots of the blockchain state at specific points in time.
+            transactions_processor_factory (Callable[[Session], TransactionsProcessor]): Creates processors to modify transactions.
+            accounts_manager_factory (Callable[[Session], AccountsManager]): Creates managers to handle account state.
+            contract_snapshot_factory (Callable[[str, Session, Transaction], ContractSnapshot]): Creates snapshots of contract states.
+            node_factory (Callable[[dict, ExecutionMode, ContractSnapshot, Receipt | None, MessageHandler, Callable[[str], ContractSnapshot]], Node]): Creates node instances that can execute contracts and process transactions.
+            stop_event (threading.Event): Control signal to terminate the appeal window process.
+        """
+        # Set a new event loop for the appeal window
+        asyncio.set_event_loop(asyncio.new_event_loop())
+
+        while not stop_event.is_set():
+            try:
+                async with asyncio.TaskGroup() as tg:
+                    with self.get_session() as session:
+                        # Get the accepted and undetermined transactions per contract address
+                        chain_snapshot = chain_snapshot_factory(session)
+                        accepted_undetermined_transactions = (
+                            chain_snapshot.get_accepted_undetermined_transactions()
+                        )
+
+                        # Iterate over the contracts
+                        for (
+                            accepted_undetermined_queue
+                        ) in accepted_undetermined_transactions.values():
+
+                            # Create a new session for each task so tasks can be run in parallel
+                            with self.get_session() as task_session:
+
+                                async def exec_appeal_window_with_session_handling(
+                                    task_session: Session,
+                                    accepted_undetermined_queue: list[dict],
+                                ):
+                                    transactions_processor = (
+                                        transactions_processor_factory(task_session)
+                                    )
+
+                                    # Go through the whole queue to check for appeals and finalizations
+                                    for index, transaction in enumerate(
+                                        accepted_undetermined_queue
+                                    ):
+                                        transaction = Transaction.from_dict(transaction)
+
+                                        # Check if the transaction is appealed
+                                        if not transaction.appealed:
+
+                                            # Check if the transaction can be finalized
+                                            if self.can_finalize_transaction(
+                                                transactions_processor,
+                                                transaction,
+                                                index,
+                                                accepted_undetermined_queue,
+                                            ):
+
+                                                # Handle transactions that need to be finalized
+                                                await self.process_finalization(
+                                                    transaction,
+                                                    transactions_processor,
+                                                    chain_snapshot,
+                                                    accounts_manager_factory(
+                                                        task_session
+                                                    ),
+                                                    lambda contract_address: contract_snapshot_factory(
+                                                        contract_address,
+                                                        task_session,
+                                                        transaction,
+                                                    ),
+                                                    node_factory,
+                                                )
+                                                task_session.commit()
+
+                                        else:
+                                            # Handle transactions that are appealed
+                                            if (
+                                                transaction.status
+                                                == TransactionStatus.UNDETERMINED
+                                            ):
+                                                # Leader appeal
+                                                await self.process_leader_appeal(
+                                                    transaction,
+                                                    transactions_processor,
+                                                    chain_snapshot,
+                                                    accounts_manager_factory(
+                                                        task_session
+                                                    ),
+                                                    lambda contract_address: contract_snapshot_factory(
+                                                        contract_address,
+                                                        task_session,
+                                                        transaction,
+                                                    ),
+                                                    node_factory,
+                                                )
+                                                task_session.commit()
+
+                                            else:
+                                                # Validator appeal
+                                                await self.process_validator_appeal(
+                                                    transaction,
+                                                    transactions_processor,
+                                                    chain_snapshot,
+                                                    accounts_manager_factory(
+                                                        task_session
+                                                    ),
+                                                    lambda contract_address: contract_snapshot_factory(
+                                                        contract_address,
+                                                        task_session,
+                                                        transaction,
+                                                    ),
+                                                    node_factory,
+                                                )
+                                                task_session.commit()
+
+                                tg.create_task(
+                                    exec_appeal_window_with_session_handling(
+                                        task_session, accepted_undetermined_queue
+                                    )
+                                )
+
+            except Exception as e:
+                print("Error running consensus", e)
+                print(traceback.format_exc())
+            await asyncio.sleep(self.consensus_sleep_time)
+
+    def can_finalize_transaction(
+        self,
+        transactions_processor: TransactionsProcessor,
+        transaction: Transaction,
+        index: int,
+        accepted_undetermined_queue: list[dict],
+    ) -> bool:
+        """
+        Check if the transaction can be finalized based on the following criteria:
+        - The transaction is a leader only transaction
+        - The transaction has exceeded the finality window
+        - The previous transaction has been finalized
+
+        Args:
+            transactions_processor (TransactionsProcessor): The transactions processor instance.
+            transaction (Transaction): The transaction to be possibly finalized.
+            index (int): The index of the current transaction in the accepted_undetermined_queue.
+            accepted_undetermined_queue (list[dict]): The list of accepted and undetermined transactions for one contract.
+
+        Returns:
+            bool: True if the transaction can be finalized, False otherwise.
+        """
+        if (transaction.leader_only) or (
+            (
+                int(time.time())
+                - transaction.timestamp_awaiting_finalization
+                - transaction.appeal_processing_time
+            )
+            > self.finality_window_time
+        ):
+            if index == 0:
+                return True
+            else:
+                previous_transaction_hash = accepted_undetermined_queue[index - 1][
+                    "hash"
+                ]
+                previous_transaction = transactions_processor.get_transaction_by_hash(
+                    previous_transaction_hash
+                )
+                if previous_transaction["status"] == TransactionStatus.FINALIZED.value:
+                    return True
+                else:
+                    return False
+        else:
+            return False
+
+    async def process_finalization(
+        self,
+        transaction: Transaction,
+        transactions_processor: TransactionsProcessor,
+        chain_snapshot: ChainSnapshot,
+        accounts_manager: AccountsManager,
+        contract_snapshot_factory: Callable[[str], ContractSnapshot],
+        node_factory: Callable[
+            [
+                dict,
+                ExecutionMode,
+                ContractSnapshot,
+                Receipt | None,
+                MessageHandler,
+                Callable[[str], ContractSnapshot],
+            ],
+            Node,
+        ],
+    ):
+        """
+        Process the finalization of a transaction.
+
+        Args:
+            transaction (Transaction): The transaction to finalize.
+            transactions_processor (TransactionsProcessor): Instance responsible for handling transaction operations within the database.
+            chain_snapshot (ChainSnapshot): Snapshot of the chain state.
+            accounts_manager (AccountsManager): Manager for accounts.
+            contract_snapshot_factory (Callable[[str], ContractSnapshot]): Factory function to create contract snapshots.
+            node_factory (Callable[[dict, ExecutionMode, ContractSnapshot, Receipt | None, MessageHandler, Callable[[str], ContractSnapshot]], Node]): Factory function to create nodes.
+        """
+        # Create a transaction context for finalizing the transaction
+        context = TransactionContext(
+            transaction=transaction,
+            transactions_processor=transactions_processor,
+            chain_snapshot=chain_snapshot,
+            accounts_manager=accounts_manager,
+            contract_snapshot_factory=contract_snapshot_factory,
+            node_factory=node_factory,
+            msg_handler=self.msg_handler,
+        )
+
+        # Transition to the FinalizingState
+        state = FinalizingState()
+        await state.handle(context)
+
+    async def process_leader_appeal(
+        self,
+        transaction: Transaction,
+        transactions_processor: TransactionsProcessor,
+        chain_snapshot: ChainSnapshot,
+        accounts_manager: AccountsManager,
+        contract_snapshot_factory: Callable[[str], ContractSnapshot],
+        node_factory: Callable[
+            [
+                dict,
+                ExecutionMode,
+                ContractSnapshot,
+                Receipt | None,
+                MessageHandler,
+                Callable[[str], ContractSnapshot],
+            ],
+            Node,
+        ],
+    ):
+        """
+        Process the leader appeal of a transaction.
+
+        Args:
+            transaction (Transaction): The transaction to appeal.
+            transactions_processor (TransactionsProcessor): Instance responsible for handling transaction operations within the database.
+            chain_snapshot (ChainSnapshot): Snapshot of the chain state.
+            accounts_manager (AccountsManager): Manager for accounts.
+            contract_snapshot_factory (Callable[[str], ContractSnapshot]): Factory function to create contract snapshots.
+            node_factory (Callable[[dict, ExecutionMode, ContractSnapshot, Receipt | None, MessageHandler, Callable[[str], ContractSnapshot]], Node]): Factory function to create nodes.
+        """
+        # Create a transaction context for the appeal
+        context = TransactionContext(
+            transaction=transaction,
+            transactions_processor=transactions_processor,
+            chain_snapshot=chain_snapshot,
+            accounts_manager=accounts_manager,
+            contract_snapshot_factory=contract_snapshot_factory,
+            node_factory=node_factory,
+            msg_handler=self.msg_handler,
+        )
+
+        transactions_processor.set_transaction_appeal(transaction.hash, False)
+        transaction.appealed = False
+
+        used_leader_addresses = (
+            ConsensusAlgorithm.get_used_leader_addresses_from_consensus_history(
+                context.transactions_processor.get_transaction_by_hash(
+                    context.transaction.hash
+                )["consensus_history"]
+            )
+        )
+        if len(transaction.consensus_data.validators) + len(
+            used_leader_addresses
+        ) >= len(chain_snapshot.get_all_validators()):
+            self.msg_handler.send_message(
+                LogEvent(
+                    "consensus_event",
+                    EventType.ERROR,
+                    EventScope.CONSENSUS,
+                    "Appeal failed, no validators found to process the appeal",
+                    {
+                        "transaction_hash": transaction.hash,
+                    },
+                    transaction_hash=transaction.hash,
+                )
+            )
+            self.msg_handler.send_message(
+                log_event=LogEvent(
+                    "transaction_appeal_updated",
+                    EventType.INFO,
+                    EventScope.CONSENSUS,
+                    "Set transaction appealed",
+                    {
+                        "hash": context.transaction.hash,
+                    },
+                ),
+                log_to_terminal=False,
+            )
+
+        else:
+            # Appeal data member is used in the frontend for both types of appeals
+            # Here the type is refined based on the status
+            transactions_processor.set_transaction_appeal_undetermined(
+                transaction.hash, True
+            )
+            transaction.appeal_undetermined = True
+
+            # Begin state transitions starting from PendingState
+            state = PendingState()
+            while True:
+                next_state = await state.handle(context)
+                if next_state is None:
+                    break
+                elif next_state == "leader_appeal_success":
+                    self.rollback_transactions(context)
+                    break
+                state = next_state
+
+    async def process_validator_appeal(
+        self,
+        transaction: Transaction,
+        transactions_processor: TransactionsProcessor,
+        chain_snapshot: ChainSnapshot,
+        accounts_manager: AccountsManager,
+        contract_snapshot_factory: Callable[[str], ContractSnapshot],
+        node_factory: Callable[
+            [
+                dict,
+                ExecutionMode,
+                ContractSnapshot,
+                Receipt | None,
+                MessageHandler,
+                Callable[[str], ContractSnapshot],
+            ],
+            Node,
+        ],
+    ):
+        """
+        Process the validator appeal of a transaction.
+
+        Args:
+            transaction (Transaction): The transaction to appeal.
+            transactions_processor (TransactionsProcessor): Instance responsible for handling transaction operations within the database.
+            chain_snapshot (ChainSnapshot): Snapshot of the chain state.
+            accounts_manager (AccountsManager): Manager for accounts.
+            contract_snapshot_factory (Callable[[str], ContractSnapshot]): Factory function to create contract snapshots.
+            node_factory (Callable[[dict, ExecutionMode, ContractSnapshot, Receipt | None, MessageHandler, Callable[[str], ContractSnapshot]], Node]): Factory function to create nodes.
+        """
+        # Create a transaction context for the appeal
+        context = TransactionContext(
+            transaction=transaction,
+            transactions_processor=transactions_processor,
+            chain_snapshot=chain_snapshot,
+            accounts_manager=accounts_manager,
+            contract_snapshot_factory=contract_snapshot_factory,
+            node_factory=node_factory,
+            msg_handler=self.msg_handler,
+        )
+
+        # Set the leader receipt in the context
+        context.consensus_data.leader_receipt = (
+            transaction.consensus_data.leader_receipt
+        )
+        try:
+            # Attempt to get extra validators for the appeal process
+            _, context.remaining_validators = ConsensusAlgorithm.get_extra_validators(
+                chain_snapshot.get_all_validators(),
+                transaction.consensus_history,
+                transaction.consensus_data,
+                transaction.appeal_failed,
+            )
+        except ValueError as e:
+            # When no validators are found, then the appeal failed
+            context.msg_handler.send_message(
+                LogEvent(
+                    "consensus_event",
+                    EventType.ERROR,
+                    EventScope.CONSENSUS,
+                    "Appeal failed, no validators found to process the appeal",
+                    {
+                        "transaction_hash": context.transaction.hash,
+                    },
+                    transaction_hash=context.transaction.hash,
+                )
+            )
+            context.transactions_processor.set_transaction_appeal(
+                context.transaction.hash, False
+            )
+            context.transaction.appealed = False
+            self.msg_handler.send_message(
+                log_event=LogEvent(
+                    "transaction_appeal_updated",
+                    EventType.INFO,
+                    EventScope.CONSENSUS,
+                    "Set transaction appealed",
+                    {
+                        "hash": context.transaction.hash,
+                    },
+                ),
+                log_to_terminal=False,
+            )
+            context.transactions_processor.set_transaction_appeal_processing_time(
+                context.transaction.hash
+            )
+        else:
+            # Set up the context for the committing state
+            context.num_validators = len(context.remaining_validators)
+            context.votes = {}
+            context.contract_snapshot_supplier = (
+                lambda: context.contract_snapshot_factory(
+                    context.transaction.to_address
+                )
+            )
+
+            # Begin state transitions starting from CommittingState
+            state = CommittingState()
+            while True:
+                next_state = await state.handle(context)
+                if next_state is None:
+                    break
+                elif next_state == "validator_appeal_success":
+                    self.rollback_transactions(context)
+                    ConsensusAlgorithm.dispatch_transaction_status_update(
+                        context.transactions_processor,
+                        context.transaction.hash,
+                        TransactionStatus.PENDING,
+                        context.msg_handler,
+                    )
+
+                    # Get the previous state of the contract
+                    previous_contact_state = (
+                        context.transaction.contract_snapshot.encoded_state
+                    )
+
+                    # Restore the contract state
+                    if previous_contact_state:
+                        # Get the contract snapshot for the transaction's target address
+                        leaders_contract_snapshot = context.contract_snapshot_factory(
+                            context.transaction.to_address
+                        )
+
+                        # Update the contract state with the previous state
+                        leaders_contract_snapshot.update_contract_state(
+                            accepted_state=previous_contact_state
+                        )
+
+                    # Transaction will be picked up by _crawl_snapshot
+                    break
+                state = next_state
+
+    def rollback_transactions(self, context: TransactionContext):
+        """
+        Rollback newer transactions.
+        """
+        # Rollback all future transactions for the current contract
+        # Stop the _crawl_snapshot and the _run_consensus for the current contract
+        address = context.transaction.to_address
+        self.stop_pending_queue_task(address)
+
+        # Wait until task is finished
+        while self.is_pending_queue_task_running(address):
+            time.sleep(1)
+
+        # Empty the pending queue
+        self.pending_queues[address] = asyncio.Queue()
+
+        # Set all transactions with higher created_at to PENDING
+        future_transactions = context.transactions_processor.get_newer_transactions(
+            context.transaction.hash
+        )
+        for future_transaction in future_transactions:
+            ConsensusAlgorithm.dispatch_transaction_status_update(
+                context.transactions_processor,
+                future_transaction["hash"],
+                TransactionStatus.PENDING,
+                context.msg_handler,
+            )
+
+            # Reset the contract snapshot for the transaction
+            context.transactions_processor.set_transaction_contract_snapshot(
+                future_transaction["hash"], None
+            )
+
+        # Start the queue loop again
+        self.start_pending_queue_task(address)
+
+    @staticmethod
+    def get_extra_validators(
+        all_validators: List[dict],
+        consensus_history: dict,
+        consensus_data: ConsensusData,
+        appeal_failed: int,
+    ):
+        """
+        Get extra validators for the appeal process according to the following formula:
+        - when appeal_failed = 0, add n + 2 validators
+        - when appeal_failed > 0, add (2 * appeal_failed * n + 1) + 2 validators
+        Note that for appeal_failed > 0, the returned set contains the old validators
+        from the previous appeal round and new validators.
+
+        Selection of the extra validators:
+        appeal_failed | PendingState | Reused validators | Extra selected     | Total
+                      | validators   | from the previous | validators for the | validators
+                      |              | appeal round      | appeal             |
+        ----------------------------------------------------------------------------------
+               0      |       n      |          0        |        n+2         |    2n+2
+               1      |       n      |        n+2        |        n+1         |    3n+3
+               2      |       n      |       2n+3        |         2n         |    5n+3
+               3      |       n      |       4n+3        |         2n         |    7n+3
+                              └───────┬──────┘  └─────────┬────────┘
+                                      │                   |
+        Validators after the ◄────────┘                   └──► Validators during the appeal
+        appeal. This equals                                    for appeal_failed > 0
+        the Total validators                                   = (2*appeal_failed*n+1)+2
+        of the row above,                                      This is the formula from
+        and are in consensus_data.                             above and it is what is
+        For appeal_failed > 0                                  returned by this function
+        = (2*appeal_failed-1)*n+3
+        This is used to calculate n
+
+        Args:
+            all_validators (List[dict]): List of all validators.
+            consensus_history (dict): Dictionary of consensus rounds results and status changes.
+            consensus_data (ConsensusData): Data related to the consensus process.
+            appeal_failed (int): Number of times the appeal has failed.
+
+        Returns:
+            list: List of current validators.
+            list: List of extra validators.
+        """
+        # Get current validators and a dictionary mapping addresses to validators not used in the consensus process
+        current_validators, validator_map = (
+            ConsensusAlgorithm.get_validators_from_consensus_data(
+                all_validators, consensus_data, False
+            )
+        )
+
+        # Remove used leaders from validator_map
+        used_leader_addresses = (
+            ConsensusAlgorithm.get_used_leader_addresses_from_consensus_history(
+                consensus_history
+            )
+        )
+        for used_leader_address in used_leader_addresses:
+            if used_leader_address in validator_map:
+                validator_map.pop(used_leader_address)
+
+        # Set not_used_validators to the remaining validators in validator_map
+        not_used_validators = list(validator_map.values())
+
+        if len(not_used_validators) == 0:
+            raise ValueError("No validators found")
+
+        nb_current_validators = len(current_validators) + 1  # including the leader
+        if appeal_failed == 0:
+            # Calculate extra validators when no appeal has failed
+            extra_validators = get_validators_for_transaction(
+                not_used_validators, nb_current_validators + 2
+            )
+        elif appeal_failed == 1:
+            # Calculate extra validators when one appeal has failed
+            n = (nb_current_validators - 2) // 2
+            extra_validators = get_validators_for_transaction(
+                not_used_validators, n + 1
+            )
+            extra_validators = current_validators[n - 1 :] + extra_validators
+        else:
+            # Calculate extra validators when more than one appeal has failed
+            n = (nb_current_validators - 3) // (2 * appeal_failed - 1)
+            extra_validators = get_validators_for_transaction(
+                not_used_validators, 2 * n
+            )
+            extra_validators = current_validators[n - 1 :] + extra_validators
+
+        return current_validators, extra_validators
+
+    @staticmethod
+    def get_validators_from_consensus_data(
+        all_validators: List[dict], consensus_data: ConsensusData, include_leader: bool
+    ):
+        """
+        Get validators from consensus data.
+
+        Args:
+            all_validators (List[dict]): List of all validators.
+            consensus_data (ConsensusData): Data related to the consensus process.
+            include_leader (bool): Whether to get the leader in the validator set.
+        Returns:
+            list: List of validators involved in the consensus process (can include the leader).
+            dict: Dictionary mapping addresses to validators not used in the consensus process.
+        """
+        # Create a dictionary to map addresses to a validator
+        validator_map = {
+            validator["address"]: validator for validator in all_validators
+        }
+
+        # Extract address of the leader from consensus data
+        if include_leader:
+            receipt_addresses = [consensus_data.leader_receipt.node_config["address"]]
+        else:
+            receipt_addresses = []
+
+        # Extract addresses of validators from consensus data
+        receipt_addresses += [
+            receipt.node_config["address"] for receipt in consensus_data.validators
+        ]
+
+        # Return validators whose addresses are in the receipt addresses
+        validators = [
+            validator_map.pop(receipt_address)
+            for receipt_address in receipt_addresses
+            if receipt_address in validator_map
+        ]
+
+        return validators, validator_map
+
+    @staticmethod
+    def add_new_validator(
+        all_validators: List[dict], validators: List[dict], leader_addresses: set[str]
+    ):
+        """
+        Add a new validator to the list of validators.
+
+        Args:
+            all_validators (List[dict]): List of all validators.
+            validators (list[dict]): List of validators.
+            leader_addresses (set[str]): Set of leader addresses.
+
+        Returns:
+            list: List of validators.
+        """
+        # Check if there is a validator to be possibly selected
+        if len(leader_addresses) + len(validators) >= len(all_validators):
+            raise ValueError("No more validators found to add a new validator")
+
+        # Extract a set of addresses of validators and leaders
+        addresses = {validator["address"] for validator in validators}
+        addresses.update(leader_addresses)
+
+        # Get not used validators
+        not_used_validators = [
+            validator
+            for validator in all_validators
+            if validator["address"] not in addresses
+        ]
+
+        # Get new validator
+        new_validator = get_validators_for_transaction(not_used_validators, 1)
+
+        return new_validator + validators
+
+    @staticmethod
+    def get_used_leader_addresses_from_consensus_history(
+        consensus_history: dict, current_leader_receipt: Receipt | None = None
+    ):
+        """
+        Get the used leader addresses from the consensus history.
+
+        Args:
+            consensus_history (dict): Dictionary of consensus rounds results and status changes.
+            current_leader_receipt (Receipt | None): Current leader receipt.
+
+        Returns:
+            set[str]: Set of used leader addresses.
+        """
+        used_leader_addresses = set()
+        if "consensus_results" in consensus_history:
+            for consensus_round in consensus_history["consensus_results"]:
+                leader_receipt = consensus_round["leader_result"]
+                if leader_receipt:
+                    used_leader_addresses.update(
+                        [leader_receipt["node_config"]["address"]]
+                    )
+
+        # consensus_history does not contain the latest consensus_data
+        if current_leader_receipt:
+            used_leader_addresses.update(
+                [current_leader_receipt.node_config["address"]]
+            )
+
+        return used_leader_addresses
+
+    def set_finality_window_time(self, time: int):
+        """
+        Set the finality window time.
+
+        Args:
+            time (int): The finality window time.
+        """
+        self.finality_window_time = time
+
+        # Send log event to update the frontend value
+        self.msg_handler.send_message(
+            LogEvent(
+                name="finality_window_time_updated",
+                type=EventType.INFO,
+                scope=EventScope.RPC,
+                message=f"Finality window time updated to {time}",
+                data={"time": time},
+            ),
+            log_to_terminal=False,
+        )
