+++ backend/consensus/states/committing_state.py
@@ -1,0 +1,71 @@
+import asyncio
+from backend.database_handler.transactions_processor import TransactionStatus
+from backend.node.base import Node
+from backend.node.types import (
+    ExecutionMode,
+    Receipt,
+)
+from backend.consensus.base import TransactionState
+from backend.consensus.states.revealing_state import RevealingState
+
+
+class CommittingState(TransactionState):
+    """
+    Class representing the committing state of a transaction.
+    """
+
+    async def handle(self, context):
+        """
+        Handle the committing state transition. There are no encrypted votes.
+
+        Args:
+            context (TransactionContext): The context of the transaction.
+
+        Returns:
+            TransactionState: The RevealingState.
+        """
+        from backend.consensus.helpers.consensus_algorithm import ConsensusAlgorithm
+
+        # Dispatch a transaction status update to COMMITTING
+        ConsensusAlgorithm.dispatch_transaction_status_update(
+            context.transactions_processor,
+            context.transaction.hash,
+            TransactionStatus.COMMITTING,
+            context.msg_handler,
+        )
+
+        context.transactions_processor.create_rollup_transaction(
+            context.transaction.hash
+        )
+
+        # Create validator nodes for each validator
+        context.validator_nodes = [
+            context.node_factory(
+                validator,
+                ExecutionMode.VALIDATOR,
+                (
+                    context.transaction.contract_snapshot
+                    if context.transaction.contract_snapshot
+                    else context.contract_snapshot_supplier()
+                ),
+                context.consensus_data.leader_receipt,
+                context.msg_handler,
+                context.contract_snapshot_factory,
+            )
+            for validator in context.remaining_validators
+        ]
+
+        # Execute the transaction on each validator node and gather the results
+        sem = asyncio.Semaphore(8)
+
+        async def run_single_validator(validator: Node) -> Receipt:
+            async with sem:
+                return await validator.exec_transaction(context.transaction)
+
+        validation_tasks = [
+            run_single_validator(validator) for validator in context.validator_nodes
+        ]
+        context.validation_results = await asyncio.gather(*validation_tasks)
+
+        # Transition to the RevealingState
+        return RevealingState()
