+++ backend/consensus/states/finalizing_state.py
@@ -1,0 +1,56 @@
+from backend.database_handler.transactions_processor import TransactionStatus
+from backend.node.types import ExecutionResultStatus
+from backend.consensus.base import TransactionState, _emit_transactions
+
+
+class FinalizingState(TransactionState):
+    """
+    Class representing the finalizing state of a transaction.
+    """
+
+    async def handle(self, context):
+        """
+        Handle the finalizing state transition.
+
+        Args:
+            context (TransactionContext): The context of the transaction.
+
+        Returns:
+            None: The transaction is finalized.
+        """
+        from backend.consensus.helpers.consensus_algorithm import ConsensusAlgorithm
+
+        # Retrieve the leader's receipt from the consensus data
+        leader_receipt = context.transaction.consensus_data.leader_receipt
+
+        # Update contract state
+        if (context.transaction.status == TransactionStatus.ACCEPTED) and (
+            leader_receipt.execution_result == ExecutionResultStatus.SUCCESS
+        ):
+            # Get the contract snapshot for the transaction's target address
+            leaders_contract_snapshot = context.contract_snapshot_factory(
+                context.transaction.to_address
+            )
+            leaders_contract_snapshot.update_contract_state(
+                finalized_state=leader_receipt.contract_state
+            )
+
+        # Update the transaction status to FINALIZED
+        ConsensusAlgorithm.dispatch_transaction_status_update(
+            context.transactions_processor,
+            context.transaction.hash,
+            TransactionStatus.FINALIZED,
+            context.msg_handler,
+        )
+
+        context.transactions_processor.create_rollup_transaction(
+            context.transaction.hash
+        )
+
+        if context.transaction.status != TransactionStatus.UNDETERMINED:
+            # Insert pending transactions generated by contract-to-contract calls
+            _emit_transactions(
+                context,
+                context.transaction.consensus_data.leader_receipt.pending_transactions,
+                "finalized",
+            )
